[CP-3213] `import-contacts` catalog model renamed to `unified-entities`
