+++ libs/device/models/src/index.ts
@@ -16,6 +16,6 @@ export * from "./lib/backup"
 export * from "./lib/file-transfer"
 export * from "./lib/restore"
 export * from "./lib/feature"
-export * from "./lib/import-contacts"
+export * from "./lib/unified-entities"
 export * from "./lib/data-transfer"
 export * from "./lib/entities"
