+++ libs/device/models/src/lib/unified-entities/index.ts
